Added a way to use a mockserver to test json respnses
The test code should just be testing the processing of specific json
returned. Since the data is dynamic there is no reason to be testing it
the production data set. So created a mock server to facilitate just
having static json responses in the dest directory instead of making
connection to the production server.